requirements.txt
torch
torchvision
numpy
einops
ptflops
pyyaml
opencv-python
albumentations